+++ hata/ext/extension_loader/utils.py
@@ -1,9 +1,10 @@
 __all__ = ()
 
 from sys import path as route_paths
-from os.path import join as join_path, isdir as is_directory, isfile as is_file
+from os.path import join as join_path, isdir as is_folder, isfile as is_file, exists
 from os import listdir as list_directory
 
+from ...backend.utils import HybridValueDictionary
 from ...backend.analyzer import CallableAnalyzer
 
 
@@ -49,6 +50,92 @@ def _validate_entry_or_exit(point):
     
     return False
 
+def validate_extension_parameters(entry_point=None, exit_point=None, extend_default_variables=True, locked=False,
+        take_snapshot_difference=True, **variables):
+    """
+    Validates extension parameters.
+    
+    Parameters
+    ----------
+    entry_point : `None`, `str` or `callable`, Optional
+        Extension specific entry point, to use over the extension loader's default.
+    exit_point : `None`, `str` or `callable`, Optional
+        Extension specific exit point, to use over the extension loader's default.
+    locked : `bool`, Optional
+        Whether the given extension(s) should not be affected by `.{}_all` methods.
+    take_snapshot_difference : `bool`, Optional
+        Whether snapshot feature should be used.
+    **variables : Keyword parameters
+        Variables to assign to an extension(s)'s module before they are loaded.
+    
+    Raises
+    ------
+    TypeError
+        - If `entry_point` was not given as `None`, `str` or as `callable`.
+        - If `entry_point` was given as `callable`, but accepts less or more positional arguments, as would be
+            given.
+        - If `exit_point` was not given as `None`, `str` or as `callable`.
+        - If `exit_point` was given as `callable`, but accepts less or more positional arguments, as would be
+            given.
+        - If `extend_default_variables` was not given as `bool`.
+        - If `locked` was not given as `bool`.
+        - If `name` was not given as `str` or as `iterable` of `str`.
+    ValueError
+        If a variable name is would be used, what is `module` attribute.
+    
+    Returns
+    -------
+    entry_point : `None`, `str` or `callable`
+        Extension specific entry point, to use over the extension loader's default.
+    exit_point : `None`, `str` or `callable`
+        Extension specific exit point, to use over the extension loader's default.
+    extend_default_variables : `bool`
+        Whether the extension should use the loader's default variables or just it's own's.
+    locked : `bool`
+        Whether the given extension(s) should not be affected by `.{}_all` methods.
+    take_snapshot_difference : `bool`
+        Whether snapshot feature should be used.
+    default_variables : `None` or `HybridValueDictionary` of (`str`, `Any`) items
+        An optionally weak value dictionary to store objects for assigning them to modules before loading them.
+        If would be empty, is set as `None` instead.
+    """
+    if not _validate_entry_or_exit(entry_point):
+        raise TypeError(f'`validate_extension_parameters` expected `None`, `str` or a `callable` as '
+            f'`entry_point`, got {entry_point.__class__.__name__}.')
+    
+    if not _validate_entry_or_exit(exit_point):
+        raise TypeError(f'`validate_extension_parameters` expected `None`, `str` or a `callable` as `exit_point`, '
+            f'got {exit_point.__class__.__name__}.')
+    
+    if variables:
+        default_variables = HybridValueDictionary(variables)
+        for key, value in variables.items():
+            if key in PROTECTED_NAMES:
+                raise ValueError(f'The passed {key!r} is a protected variable name of module type.')
+            default_variables[key] = value
+    else:
+        default_variables = None
+    
+    extend_default_variables_type = extend_default_variables.__class__
+    if extend_default_variables_type is bool:
+        pass
+    elif issubclass(extend_default_variables_type, int):
+        extend_default_variables = bool(extend_default_variables)
+    else:
+        raise TypeError(f'`extend_default_variables` should have been passed as `bool`, got: '
+            f'{extend_default_variables_type.__name__}.')
+    
+    locked_type = type(locked)
+    if locked_type is bool:
+        pass
+    elif issubclass(locked_type, int):
+        locked = bool(locked)
+    else:
+        raise TypeError(f'`locked` should have been passed as `bool`, got: {locked_type.__name__}.')
+    
+    return entry_point, exit_point, extend_default_variables, locked, take_snapshot_difference, default_variables
+
+
 PROTECTED_NAMES = {
     '__class__',
     '__delattr__',
@@ -89,6 +176,15 @@ PROTECTED_NAMES = {
 }
 
 
+
+PYTHON_EXTENSION_NAMES = (
+    '.py',
+    '.pyd',
+    '.pyc',
+    '.so',
+)
+
+
 def _iter_extension_names(name):
     """
     Fetches the names
@@ -107,11 +203,13 @@ def _iter_extension_names(name):
     
     Raises
     ------
+    ImportError
+        If a name could not be detected as an extension.
     TypeError
         If `name` is not `str` nor an `iterable` of `str` instances.
     """
     for name in _iter_name_maybe_iterable(name):
-        yield from _iter_name_maybe_directory(name)
+        yield from _lookup_path(name)
 
 
 def _iter_name_maybe_iterable(name):
@@ -155,37 +253,82 @@ def _iter_name_maybe_iterable(name):
             f'{name_type.__name__}.')
 
 
-def _iter_name_maybe_directory(name):
+def _lookup_path(import_name):
     """
-    Iterates over the extension's names if it is a directory.
+    Detects the root of the given name.
     
-    This method is a generator.
+    This function is an iterable generator.
     
     Parameters
     ----------
-    name : `str`
-        An extension's name.
+    import_name : `str`
+        An extension's import name.
     
     Yields
     ------
-    name : `str`
-        Extension names.
+    import_name : `str`
+        Import name to an extension file.
+    
+    Raise
+    -----
+    ImportError
+        If `name` name could not be detected as en extension.
     """
-    path_end = join_path(*name.split('.'))
+    path_end = join_path(*import_name.split('.'))
     for base_path in route_paths:
         path = join_path(base_path, path_end)
-        if is_directory(path) and (not is_file(join_path(path, '__init__.py'))):
-            for file_name in list_directory(path):
-                file_path = join_path(path, file_name)
-                if is_directory(file_path):
-                    if is_file(join_path(path, '__init__.py')):
-                        yield f'{name}.{file_name}'
-                    continue
-                
-                if is_file(file_path):
-                    if file_name.endswith('.py'):
-                        yield f'{name}.{file_name[:-3]}'
-                    continue
+        if exists(path) and is_folder(path):
+            yield from _iter_folder(import_name, path)
+            return
+        
+        for python_extension_name in PYTHON_EXTENSION_NAMES:
+            file_path = path+python_extension_name
+            if exists(file_path) and is_file(file_path):
+                yield import_name
+                return
+    
+    raise TypeError(f'The given `import_name` could nto be detected as an extension, got {import_name!r}.')
+
+
+def _iter_folder(import_name, folder_path):
+    """
+    Iterates over a folder's import names.
+    Parameters
+    ----------
+    import_name : `str`
+        The name of the extension if we would import it.
+    folder_path : `str`
+        Path to the folder
+    
+    Yields
+    ------
+    import_name : `str`
+        Detected import names for each applicable file in the folder.
+    """
+    for python_extension_name in PYTHON_EXTENSION_NAMES:
+        file_path = join_path(folder_path, f'__init__{python_extension_name}')
+        if exists(file_path) and is_file(file_path):
+            yield import_name
             return
     
-    yield name
+    for file_name in list_directory(folder_path):
+        if file_name.startswith('.') or (file_name == '__pycache__'):
+            continue
+        
+        path = join_path(folder_path, file_name)
+        
+        if is_file(path):
+            for python_extension_name in PYTHON_EXTENSION_NAMES:
+                if file_name.endswith(python_extension_name):
+                    value = f'{import_name}.{file_name[:-len(python_extension_name)]}'
+                    yield value
+                    break
+            
+            continue
+        
+        if is_folder(path):
+            yield from _iter_folder(f'{import_name}.{file_name}', path)
+            continue
+        
+        # no more cases
+        continue